apps/backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Pydantic for data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver (backup)

# Redis for caching and rate limiting
redis==5.0.1
hiredis==2.2.3  # Faster Redis parser
slowapi==0.1.9  # Rate limiting for FastAPI
schedule==1.2.0  # Background job scheduling for payment email monitoring
apscheduler==3.10.4  # Advanced Python Scheduler for follow-ups

# Celery for async task queue (escalation notifications, recording downloads)
celery==5.3.4
kombu==5.3.4  # Celery messaging library
celery[redis]==5.3.4  # Celery with Redis broker support
flower==2.0.1  # Celery monitoring web UI (optional)
billiard==4.2.0  # Celery process pool implementation

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP client for integrations
httpx==0.25.2
aiohttp==3.9.1

# OpenAI integration
openai==1.3.8

# RingCentral integration
ringcentral==0.9.2

# Stripe integration
stripe==7.8.0

# Plaid integration
plaid-python==12.0.0

# Meta/Facebook integration
facebook-business==18.0.2

# Google Cloud integration
google-cloud-storage==2.10.0
google-cloud-secretmanager==2.18.1

# Email
email-validator==2.1.0
jinja2==3.1.2

# Environment and configuration
python-dotenv==1.0.0

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.26

# Data processing and analytics
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2  # ML for intent clustering, lead scoring
plotly==5.18.0  # Interactive charts for admin dashboard
seaborn==0.13.0  # Statistical data visualization
sentence-transformers==2.2.2  # Embeddings for semantic search

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==5.1.0
httpx==0.25.2  # For testing async clients
faker==37.11.0  # Generate realistic test data

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Performance
orjson==3.9.10  # Faster JSON serialization

# QR Code Tracking
user-agents==2.2.0  # User agent parsing for device detection

# Voice AI - Speech-to-Text and Text-to-Speech (Deepgram Only)
deepgram-sdk>=5.3.0  # STT ($0.0125/min) + TTS ($0.015/1000 chars) - All-in-one solution
pydub==0.25.1  # Audio manipulation and format conversion